Update web.php
So far, I updated the routes/api.php, /web.php and the BlogController files. I added the route names to be in sync with the changes I made in the controller. 

Bro, for the /logout route, I've changed it to a POST request so it would be easy to log users out securely (if you don't mind). The delete and edit route I added in the BlogController has been updated here. You can review.